+++ src/test/java/com/github/bgora/rpnlibrary/CalculatorTest.java
@@ -34,7 +34,7 @@ public class CalculatorTest {
     private Calculator calc;
 
     @Before
-    public void setUp() throws Exception {
+    public void setUp() {
         calc = Calculator.createCalculator();
     }
