+++ src/test/java/com/github/bgora/rpnlibrary/CalculatorTest2ParamFunction.java
@@ -32,7 +32,7 @@ public class CalculatorTest2ParamFunction {
     private Calculator calc;
 
     @Before
-    public void setUp() throws Exception {
+    public void setUp() {
         calc = Calculator.createCalculator();
     }
